+++ xc3_model/src/material.rs
@@ -95,13 +95,13 @@ impl MaterialParameters {
     }
 }
 
-/// Selects an [ImageTexture] and [Sampler](crate::Sampler).
+/// Selects an [ImageTexture] and [Sampler].
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Texture {
     /// The index of the [ImageTexture] in [image_textures](struct.ModelRoot.html#structfield.image_textures).
     pub image_texture_index: usize,
-    /// The index of the [Sampler](crate::Sampler) in [samplers](struct.ModelGroup.html#structfield.samplers).
+    /// The index of the [Sampler] in [samplers](struct.ModelGroup.html#structfield.samplers).
     pub sampler_index: usize,
 }