+++ xc3_wgpu/src/monolib.rs
@@ -2,7 +2,7 @@ use std::{collections::BTreeMap, path::Path};
 
 use crate::texture::create_texture;
 
-/// Textures and resources from the game's `monolib/shader` folder.
+/// Texture resources from the game's `monolib/shader` folder.
 pub struct MonolibShaderTextures {
     global_textures: BTreeMap<String, wgpu::Texture>,
 }
@@ -11,32 +11,18 @@ impl MonolibShaderTextures {
     pub fn from_file<P: AsRef<Path>>(device: &wgpu::Device, queue: &wgpu::Queue, path: P) -> Self {
         let textures = xc3_model::monolib::ShaderTextures::from_folder(path.as_ref());
 
-        // TODO: How to avoid duplicating this list with xc3_model?
-        let global_textures = [
-            "gTResidentTex02",
-            "gTResidentTex04",
-            "gTResidentTex06",
-            "gTResidentTex08",
-            "gTResidentTex11",
-            "gTResidentTex08",
-            "gTResidentTex09",
-            "gTResidentTex43",
-            "gTResidentTex44",
-            "gTResidentTex45",
-            "gTResidentTex46",
-            "gTToonGrad",
-            "gTToonDarkGrad",
-        ]
-        .into_iter()
-        .filter_map(|name| {
-            Some((
-                name.to_string(),
-                textures
-                    .global_texture(name)
-                    .map(|image| create_texture(device, queue, &image))?,
-            ))
-        })
-        .collect();
+        let global_textures = textures
+            .textures
+            .keys()
+            .filter_map(|name| {
+                Some((
+                    name.to_string(),
+                    textures
+                        .global_texture(name)
+                        .map(|image| create_texture(device, queue, image))?,
+                ))
+            })
+            .collect();
 
         Self { global_textures }
     }