+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improved framerates for model and map rendering.
 * Optimized shader database creation.
 * Adjusted wimdo model export for xc3_model to rebuild skinning information.
+* Changed `xc3_model::monolib::ShaderTextures` to use a map of sampler names to textures instead of a field for each texture.
+* Improved the number of textures supported by `xc3_model::monolib::ShaderTextures`.
 
 ### Fixed
 * Fixed an issue where some eyes would not render correctly due to incorrect handling of metalness.